Make the Pusher API client a shared RwLock
This makes it possible to switch out the underlying API client as
necessary.

This is especially useful when authenticating using short-lived
credentials -- if some background worker thread keeps a handle to the
API client, it is able to replace it with new credentials before the old
ones expire (as may happen during lengthy / hefty upload sessions).