attic/nix_store: update bindings build for Nix 2.26
`#include <nix/…>` is broken by the componentized Meson build. This
is a regression, but the `.pc` files add the `include/nix` directory
of each library to the include path directly, so we can work around
it by dropping the prefixes and using `system-deps` to simplify the
native dependency handling.

I’ve discussed this with Nix upstream, who plan on fixing it;
the `.hpp` part of this commit should be reverted after it is fixed
upstream, as the `.pc` files will no longer pollute the namespace in
this manner and the previous form of the `#include`s will need to
be restored. Apparently the `-include nix/config-*.hh` pattern may
change as well.